+++ pkg/application/applyBlueprintSpecUseCase_test.go
@@ -13,13 +13,17 @@ import (
 func TestNewApplyBlueprintSpecUseCase(t *testing.T) {
 	repoMock := newMockBlueprintSpecRepository(t)
 	installUseCaseMock := newMockDoguInstallationUseCase(t)
+	componentInstallUseCaseMock := newMockComponentInstallationUseCase(t)
 	healthMock := newMockEcosystemHealthUseCase(t)
+	maintenanceMock := newMockMaintenanceMode(t)
 
-	sut := NewApplyBlueprintSpecUseCase(repoMock, installUseCaseMock, healthMock, nil)
+	sut := NewApplyBlueprintSpecUseCase(repoMock, installUseCaseMock, healthMock, componentInstallUseCaseMock, maintenanceMock)
 
 	assert.Equal(t, installUseCaseMock, sut.doguInstallUseCase)
+	assert.Equal(t, componentInstallUseCaseMock, sut.componentInstallUseCase)
 	assert.Equal(t, repoMock, sut.repo)
 	assert.Equal(t, healthMock, sut.healthUseCase)
+	assert.Equal(t, maintenanceMock, sut.maintenanceModeAdapter)
 }
 
 func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) {
@@ -33,7 +37,7 @@ func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) 
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(nil)
 		maintenanceMock.EXPECT().Activate(mock.Anything).Return(nil)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -43,7 +47,7 @@ func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) 
 	t.Run("repo error while loading", func(t *testing.T) {
 		repoMock := newMockBlueprintSpecRepository(t)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(nil, assert.AnError)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, nil)
 
 		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -59,7 +63,7 @@ func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) 
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(assert.AnError)
 		maintenanceMock.EXPECT().Activate(mock.Anything).Return(nil)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -74,7 +78,7 @@ func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) 
 		maintenanceMock := newMockMaintenanceMode(t)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		maintenanceMock.EXPECT().Activate(mock.Anything).Return(assert.AnError)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -90,7 +94,7 @@ func TestApplyBlueprintSpecUseCase_PreProcessBlueprintApplication(t *testing.T) 
 		maintenanceMock := newMockMaintenanceMode(t)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(nil)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -117,20 +121,54 @@ func TestApplyBlueprintSpecUseCase_markInProgress(t *testing.T) {
 		assert.Equal(t, domain.StatusPhaseInProgress, spec.Status)
 	})
 
-	t.Run("repo error", func(t *testing.T) {
+	t.Run("repo error while saving", func(t *testing.T) {
 		spec := &domain.BlueprintSpec{
 			Status: domain.StatusPhaseEcosystemHealthyUpfront,
 		}
 
 		repoMock := newMockBlueprintSpecRepository(t)
+		maintenanceMock := newMockMaintenanceMode(t)
+		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(assert.AnError)
-		installUseCaseMock := newMockDoguInstallationUseCase(t)
-		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock}
+		maintenanceMock.EXPECT().Activate(mock.Anything).Return(nil)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
-		err := useCase.startApplying(testCtx, spec)
+		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
 
 		require.ErrorIs(t, err, assert.AnError)
-		assert.Equal(t, domain.StatusPhaseInProgress, spec.Status)
+	})
+	t.Run("error activating maintenance mode", func(t *testing.T) {
+		spec := &domain.BlueprintSpec{
+			Status: domain.StatusPhaseEcosystemHealthyUpfront,
+		}
+
+		repoMock := newMockBlueprintSpecRepository(t)
+		maintenanceMock := newMockMaintenanceMode(t)
+		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
+		maintenanceMock.EXPECT().Activate(mock.Anything).Return(assert.AnError)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
+
+		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
+
+		require.ErrorIs(t, err, assert.AnError)
+	})
+	t.Run("do nothing on dry run", func(t *testing.T) {
+		spec := &domain.BlueprintSpec{
+			Status: domain.StatusPhaseEcosystemHealthyUpfront,
+			Config: domain.BlueprintConfiguration{DryRun: true},
+		}
+
+		repoMock := newMockBlueprintSpecRepository(t)
+		maintenanceMock := newMockMaintenanceMode(t)
+		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
+		repoMock.EXPECT().Update(testCtx, spec).Return(nil)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
+
+		err := useCase.PreProcessBlueprintApplication(testCtx, blueprintId)
+
+		require.NoError(t, err)
+		require.Equal(t, 1, len(spec.Events))
+		assert.Equal(t, domain.BlueprintDryRunEvent{}, spec.Events[0])
 	})
 }
 
@@ -222,7 +260,11 @@ func TestApplyBlueprintSpecUseCase_ApplyBlueprintSpec(t *testing.T) {
 
 		installUseCaseMock := newMockDoguInstallationUseCase(t)
 		installUseCaseMock.EXPECT().ApplyDoguStates(testCtx, "blueprintId").Return(nil)
-		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock}
+		componentInstallUseCase := newMockComponentInstallationUseCase(t)
+		componentInstallUseCase.EXPECT().ApplyComponentStates(testCtx, "blueprintId").Return(nil)
+		componentInstallUseCase.EXPECT().WaitForHealthyComponents(testCtx).Return(ecosystem.ComponentHealthResult{}, nil)
+
+		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock, componentInstallUseCase: componentInstallUseCase}
 
 		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
 
@@ -250,8 +292,6 @@ func TestApplyBlueprintSpecUseCase_ApplyBlueprintSpec(t *testing.T) {
 		repoMock.EXPECT().GetById(testCtx, "blueprintId").Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(assert.AnError)
 
-		// installUseCaseMock := newMockDoguInstallationUseCase(t)
-		// installUseCaseMock.EXPECT().ApplyDoguStates(testCtx, "blueprintId").Return(nil)
 		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: nil}
 
 		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
@@ -260,7 +300,7 @@ func TestApplyBlueprintSpecUseCase_ApplyBlueprintSpec(t *testing.T) {
 		assert.Equal(t, domain.StatusPhaseInProgress, spec.Status)
 	})
 
-	t.Run("fail to apply state", func(t *testing.T) {
+	t.Run("fail to apply component state", func(t *testing.T) {
 		spec := &domain.BlueprintSpec{
 			Status: domain.StatusPhaseEcosystemHealthyUpfront,
 		}
@@ -268,9 +308,49 @@ func TestApplyBlueprintSpecUseCase_ApplyBlueprintSpec(t *testing.T) {
 		repoMock.EXPECT().GetById(testCtx, "blueprintId").Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(nil).Times(2)
 
+		componentInstallUseCaseMock := newMockComponentInstallationUseCase(t)
+		componentInstallUseCaseMock.EXPECT().ApplyComponentStates(testCtx, "blueprintId").Return(assert.AnError)
+		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: nil, componentInstallUseCase: componentInstallUseCaseMock}
+
+		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
+
+		require.ErrorIs(t, err, assert.AnError)
+		assert.Equal(t, domain.StatusPhaseBlueprintApplicationFailed, spec.Status)
+	})
+
+	t.Run("fail to wait for component health", func(t *testing.T) {
+		spec := &domain.BlueprintSpec{
+			Status: domain.StatusPhaseEcosystemHealthyUpfront,
+		}
+		repoMock := newMockBlueprintSpecRepository(t)
+		repoMock.EXPECT().GetById(testCtx, "blueprintId").Return(spec, nil)
+		repoMock.EXPECT().Update(testCtx, spec).Return(nil).Times(2)
+
+		componentInstallUseCaseMock := newMockComponentInstallationUseCase(t)
+		componentInstallUseCaseMock.EXPECT().ApplyComponentStates(testCtx, "blueprintId").Return(nil)
+		componentInstallUseCaseMock.EXPECT().WaitForHealthyComponents(testCtx).Return(ecosystem.ComponentHealthResult{}, assert.AnError)
+		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: nil, componentInstallUseCase: componentInstallUseCaseMock}
+
+		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
+
+		require.ErrorIs(t, err, assert.AnError)
+		assert.Equal(t, domain.StatusPhaseBlueprintApplicationFailed, spec.Status)
+	})
+
+	t.Run("fail to apply dogu state", func(t *testing.T) {
+		spec := &domain.BlueprintSpec{
+			Status: domain.StatusPhaseEcosystemHealthyUpfront,
+		}
+		repoMock := newMockBlueprintSpecRepository(t)
+		repoMock.EXPECT().GetById(testCtx, "blueprintId").Return(spec, nil)
+		repoMock.EXPECT().Update(testCtx, spec).Return(nil).Times(2)
+
+		componentInstallUseCaseMock := newMockComponentInstallationUseCase(t)
+		componentInstallUseCaseMock.EXPECT().ApplyComponentStates(testCtx, "blueprintId").Return(nil)
+		componentInstallUseCaseMock.EXPECT().WaitForHealthyComponents(testCtx).Return(ecosystem.ComponentHealthResult{}, nil)
 		installUseCaseMock := newMockDoguInstallationUseCase(t)
 		installUseCaseMock.EXPECT().ApplyDoguStates(testCtx, "blueprintId").Return(assert.AnError)
-		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock}
+		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock, componentInstallUseCase: componentInstallUseCaseMock}
 
 		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
 
@@ -294,9 +374,12 @@ func TestApplyBlueprintSpecUseCase_ApplyBlueprintSpec(t *testing.T) {
 			}
 		})
 
+		componentInstallUseCaseMock := newMockComponentInstallationUseCase(t)
+		componentInstallUseCaseMock.EXPECT().ApplyComponentStates(testCtx, "blueprintId").Return(nil)
+		componentInstallUseCaseMock.EXPECT().WaitForHealthyComponents(testCtx).Return(ecosystem.ComponentHealthResult{}, nil)
 		installUseCaseMock := newMockDoguInstallationUseCase(t)
 		installUseCaseMock.EXPECT().ApplyDoguStates(testCtx, "blueprintId").Return(assert.AnError)
-		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock}
+		useCase := ApplyBlueprintSpecUseCase{repo: repoMock, doguInstallUseCase: installUseCaseMock, componentInstallUseCase: componentInstallUseCaseMock}
 
 		err := useCase.ApplyBlueprintSpec(testCtx, "blueprintId")
 
@@ -311,7 +394,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		repoMock := newMockBlueprintSpecRepository(t)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(nil, assert.AnError)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthUpfront(testCtx, blueprintId)
@@ -329,7 +412,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().CheckEcosystemHealth(testCtx, false, false).Return(ecosystem.HealthResult{}, assert.AnError)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthUpfront(testCtx, blueprintId)
@@ -349,7 +432,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthUpfront(testCtx, blueprintId)
@@ -371,7 +454,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, true, false).Return(ecosystem.HealthResult{}, nil)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthUpfront(testCtx, blueprintId)
@@ -389,7 +472,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthUpfront(testCtx, blueprintId)
@@ -405,7 +488,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		repoMock := newMockBlueprintSpecRepository(t)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(nil, assert.AnError)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthAfterwards(testCtx, blueprintId)
@@ -424,7 +507,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx).Return(ecosystem.HealthResult{}, assert.AnError)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthAfterwards(testCtx, blueprintId)
@@ -445,7 +528,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx).Return(ecosystem.HealthResult{}, nil)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthAfterwards(testCtx, blueprintId)
@@ -466,7 +549,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		healthMock := newMockEcosystemHealthUseCase(t)
 		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx).Return(ecosystem.HealthResult{}, nil)
 
-		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
+		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil, nil)
 
 		// when
 		err := sut.CheckEcosystemHealthAfterwards(testCtx, blueprintId)
@@ -488,7 +571,7 @@ func TestApplyBlueprintSpecUseCase_PostProcessBlueprintApplication(t *testing.T)
 		maintenanceMock.EXPECT().Deactivate().Return(nil)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(nil)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PostProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -503,7 +586,7 @@ func TestApplyBlueprintSpecUseCase_PostProcessBlueprintApplication(t *testing.T)
 		maintenanceMock := newMockMaintenanceMode(t)
 		maintenanceMock.EXPECT().Deactivate().Return(nil)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(nil, assert.AnError)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PostProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -520,7 +603,7 @@ func TestApplyBlueprintSpecUseCase_PostProcessBlueprintApplication(t *testing.T)
 		maintenanceMock.EXPECT().Deactivate().Return(nil)
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(spec, nil)
 		repoMock.EXPECT().Update(testCtx, spec).Return(assert.AnError)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PostProcessBlueprintApplication(testCtx, blueprintId)
 
@@ -531,7 +614,7 @@ func TestApplyBlueprintSpecUseCase_PostProcessBlueprintApplication(t *testing.T)
 		repoMock := newMockBlueprintSpecRepository(t)
 		maintenanceMock := newMockMaintenanceMode(t)
 		maintenanceMock.EXPECT().Deactivate().Return(assert.AnError)
-		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, maintenanceMock)
+		useCase := NewApplyBlueprintSpecUseCase(repoMock, nil, nil, nil, maintenanceMock)
 
 		err := useCase.PostProcessBlueprintApplication(testCtx, blueprintId)